【-】修改参数

+++ xxl-job-admin-pg/src/main/java/org/poem/controller/JobCodeController.java
@@ -33,7 +33,7 @@ public class JobCodeController {
     private XxlJobLogGlueDao xxlJobLogGlueDao;
 
     @RequestMapping
-    public String index(HttpServletRequest request, Model model, int jobId) {
+    public String index(HttpServletRequest request, Model model, Long jobId) {
         XxlJobInfo jobInfo = xxlJobInfoDao.loadById(jobId);
         List<XxlJobLogGlue> jobLogGlues = xxlJobLogGlueDao.findByJobId(jobId);
 
@@ -57,7 +57,7 @@ public class JobCodeController {
 
     @RequestMapping("/save")
     @ResponseBody
-    public ReturnT<String> save(Model model, int id, String glueSource, String glueRemark) {
+    public ReturnT<String> save(Model model, Long id, String glueSource, String glueRemark) {
         // valid
         if (glueRemark == null) {
             return new ReturnT<String>(500, (I18nUtil.getString("system_please_input") + I18nUtil.getString("jobinfo_glue_remark")));
