+++ xxl-job-admin-pg/src/main/java/org/poem/controller/JobInfoController.java
@@ -42,7 +42,7 @@ public class JobInfoController {
     private XxlJobService xxlJobService;
 
     @RequestMapping
-    public String index(HttpServletRequest request, Model model, @RequestParam(required = false, defaultValue = "-1") int jobGroup) {
+    public String index(HttpServletRequest request, Model model, @RequestParam(required = false, defaultValue = "-1") Long jobGroup) {
 
         // 枚举-字典
         model.addAttribute("ExecutorRouteStrategyEnum", ExecutorRouteStrategyEnum.values());        // 路由策略-列表
@@ -85,7 +85,7 @@ public class JobInfoController {
         return jobGroupList;
     }
 
-    public static void validPermission(HttpServletRequest request, int jobGroup) {
+    public static void validPermission(HttpServletRequest request, Long jobGroup) {
         XxlJobUser loginUser = (XxlJobUser) request.getAttribute(LoginService.LOGIN_IDENTITY_KEY);
         if (!loginUser.validPermission(jobGroup)) {
             throw new RuntimeException(I18nUtil.getString("system_permission_limit") + "[username=" + loginUser.getUsername() + "]");
@@ -96,7 +96,7 @@ public class JobInfoController {
     @ResponseBody
     public Map<String, Object> pageList(@RequestParam(required = false, defaultValue = "0") int start,
                                         @RequestParam(required = false, defaultValue = "10") int length,
-                                        int jobGroup, int triggerStatus, String jobDesc, String executorHandler, String author) {
+                                        Long jobGroup, int triggerStatus, String jobDesc, String executorHandler, String author) {
 
         return xxlJobService.pageList(start, length, jobGroup, triggerStatus, jobDesc, executorHandler, author);
     }
@@ -115,26 +115,26 @@ public class JobInfoController {
 
     @RequestMapping("/remove")
     @ResponseBody
-    public ReturnT<String> remove(int id) {
+    public ReturnT<String> remove(Long id) {
         return xxlJobService.remove(id);
     }
 
     @RequestMapping("/stop")
     @ResponseBody
-    public ReturnT<String> pause(int id) {
+    public ReturnT<String> pause(Long id) {
         return xxlJobService.stop(id);
     }
 
     @RequestMapping("/start")
     @ResponseBody
-    public ReturnT<String> start(int id) {
+    public ReturnT<String> start(Long id) {
         return xxlJobService.start(id);
     }
 
     @RequestMapping("/trigger")
     @ResponseBody
     //@PermissionLimit(limit = false)
-    public ReturnT<String> triggerJob(int id, String executorParam, String addressList) {
+    public ReturnT<String> triggerJob(Long id, String executorParam, String addressList) {
         // force cover job param
         if (executorParam == null) {
             executorParam = "";
