+++ xxl-job-admin-pg/src/main/java/org/poem/dao/XxlJobLogGlueDao.java
@@ -16,10 +16,10 @@ public interface XxlJobLogGlueDao {
 
     public int save(XxlJobLogGlue xxlJobLogGlue);
 
-    public List<XxlJobLogGlue> findByJobId(@Param("jobId") int jobId);
+    public List<XxlJobLogGlue> findByJobId(@Param("jobId") Long jobId);
 
-    public int removeOld(@Param("jobId") long jobId, @Param("limit") int limit);
+    public int removeOld(@Param("jobId") Long jobId, @Param("limit") int limit);
 
-    public int deleteByJobId(@Param("jobId") int jobId);
+    public int deleteByJobId(@Param("jobId") Long jobId);
 
 }
