+++ xxl-job-admin-pg/src/main/java/org/poem/service/XxlJobService.java
@@ -25,7 +25,7 @@ public interface XxlJobService {
      * @param author
      * @return
      */
-    public Map<String, Object> pageList(int start, int length, int jobGroup, int triggerStatus, String jobDesc, String executorHandler, String author);
+    public Map<String, Object> pageList(int start, int length, Long jobGroup, int triggerStatus, String jobDesc, String executorHandler, String author);
 
     /**
      * add job
@@ -50,7 +50,7 @@ public interface XxlJobService {
      * @param id
      * @return
      */
-    public ReturnT<String> remove(int id);
+    public ReturnT<String> remove(Long id);
 
     /**
      * start job
@@ -58,7 +58,7 @@ public interface XxlJobService {
      * @param id
      * @return
      */
-    public ReturnT<String> start(int id);
+    public ReturnT<String> start(Long id);
 
     /**
      * stop job
@@ -66,7 +66,7 @@ public interface XxlJobService {
      * @param id
      * @return
      */
-    public ReturnT<String> stop(int id);
+    public ReturnT<String> stop(Long id);
 
     /**
      * dashboard info
